requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-cosmos
python-dotenv
openai
msal
azure-identity
azure-core
azure-storage-blob
python-dateutil
python-pptx